pyproject.toml
[tool.poetry]
name = "senjibot"
version = "0.1.0"
description = ""
authors = ["Senjienji <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
Flask = "^2.0.2"
replit = "^3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
